// src/types/Global.styled.ts
export interface GlobalStyleProps {
    justifyContent?: string;
    overflow?: string;
    alignItems?: string;
    flexDirection?: string;
    flexWrap?: string;
    flex?: string;
    width?: string;
    height?: string;
    minWidth?: string;
    minHeight?: string;
    maxWidth?: string;
    maxHeight?: string;
    margin?: string;
    padding?: string;
    border?: string;
    borderRadius?: string;
    background?: string;
    color?: string;
    fontSize?: string;
    fontWeight?: string;
    fontFamily?: string;
    lineHeight?: string;
    textAlign?: string;
    textDecoration?: string;
    mr?: number;
    ml?: number;
    mt?: number;
    mb?: number;
    pt?: number;
    pb?: number;
    pl?: number;
    pr?: number;
    mx?: number;
    my?: number;
    px?: number;
    py?: number;
}